requirements.txt
langchain
langgraph
streamlit
openai
chromadb
tiktoken
pydantic
langchain-community
langchain-google-genai
langchain-chroma
langchain-huggingface
sentence-transformers
fastapi
uvicorn
guardrails-ai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
google-generativeai
requests
mcp>=0.1.0
